requirements.txt
openai==1.93.2
python-dotenv==1.1.1
fastapi==0.116.0
uvicorn==0.35.0
pydantic==2.11.7
chromadb==1.0.15
groq==0.29.0